refactor: update rated item models and mappers
