+++ app/src/main/java/com/karrar/movieapp/ui/profile/myratings/RatedUIState.kt
@@ -9,6 +9,6 @@ data class RatedUIState(
     val releaseDate: String,
     val duration: String,
     val category: String,
-    val userRating: String = ""
+    val userRating: Int = 0
 
 )
