+++ app/src/main/java/com/karrar/movieapp/ui/profile/myratings/RatedUIStateMapper.kt
@@ -3,10 +3,10 @@ package com.karrar.movieapp.ui.profile.myratings
 import com.karrar.movieapp.domain.mappers.MapperWithGenreList
 import com.karrar.movieapp.domain.models.Genre
 import com.karrar.movieapp.domain.models.Rated
+import java.util.Locale
 import javax.inject.Inject
 
-class RatedUIStateMapper @Inject constructor()
-    : MapperWithGenreList<Rated, RatedUIState> {
+class RatedUIStateMapper @Inject constructor() : MapperWithGenreList<Rated, RatedUIState> {
 
     override fun map(input: Rated, categoryIdList: List<Genre>): RatedUIState {
         val categoryNames = input.categoryIdList
@@ -15,12 +15,12 @@ class RatedUIStateMapper @Inject constructor()
             id = input.id,
             title = input.title,
             posterPath = input.posterPath,
-            rating = input.rating.toString(),
+            rating = String.format(Locale.ENGLISH, "%.1f", input.rating),
             mediaType = input.mediaType,
             releaseDate = TimeFormatters.formatDate(input.releaseDate),
             category = categoryNames.joinToString(", "),
             duration = input.duration,
-
+            userRating = input.userRating
         )
     }
 }
